+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 v2.2.0 is a feature release with the following features, fixes and enhancements:
 
-- Added documentation with an example of FIPS compliant communication with Kafka cluster.
+ - [KIP-339](https://cwiki.apache.org/confluence/display/KAFKA/KIP-339%3A+Create+a+new+IncrementalAlterConfigs+API)
+   IncrementalAlterConfigs API (#1517).
+ - [KIP-554](https://cwiki.apache.org/confluence/display/KAFKA/KIP-554%3A+Add+Broker-side+SCRAM+Config+API):
+   User SASL/SCRAM credentials alteration and description (#1575).
+ - Added documentation with an example of FIPS compliant communication with Kafka cluster.
 
 confluent-kafka-python is based on librdkafka v2.2.0, see the
 [librdkafka release notes](https://github.com/confluentinc/librdkafka/releases/tag/v2.2.0)
@@ -14,7 +18,6 @@ for a complete list of changes, enhancements, fixes and upgrade considerations.
 ## v2.1.1
 
 v2.1.1 is a maintenance release with the following fixes and enhancements:
-
 
 ### Fixes